+++ Makefile
@@ -2,10 +2,10 @@
 
 all:
 	cd Pilaf/; make
-	cd Monkey/; make
+	cd src/; make
 	cd Redis/; make
 
 clean:
 	cd Pilaf/; make clean
-	cd Monkey/; make clean
+	cd src/; make clean
 	cd Redis/; make clean
