+++ config/pilaf.cnf
@@ -4,7 +4,7 @@
 192.168.5.17 36002
 192.168.5.18 36002
 192.168.5.21 36002
-192.168.5.22 36002
+192.168.5.20 36002
 192.168.5.23 36002
 192.168.5.24 36002
 192.168.5.25 36002
