+++ src/bitmap.cc
