+++ src/bitmap.h
