+++ src/bitmap_deamon.cc
