+++ src/check_deamon_status.sh
