+++ src/check_workers_status.sh
