+++ src/generate_bitmap.cc
