+++ src/image_search_client.cc
