+++ src/image_search_client.h
