+++ src/image_search_constants.h
