+++ src/image_search_test.cc
