+++ src/image_server_main.cc
