+++ src/integrity_check.cc
