+++ src/start_all_bmp_deamons.sh
