+++ src/stop_deamon.sh
